+++ src/InventoryM.java
@@ -420,11 +420,11 @@ public class InventoryM extends javax.swing.JPanel {
             java.util.logging.Logger.getLogger(InventoryM.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         }
 
-        float qty, price;
+        double qty, price;
         double tot;
         try {
-            qty = Float.parseFloat(this.UpdatedQty.getText());
-            price = Float.parseFloat(this.AvgPrch.getText());
+            qty = Double.parseDouble(this.UpdatedQty.getText());
+            price = Double.parseDouble(this.AvgPrch.getText());
         } catch (NumberFormatException e) {
             e.printStackTrace();
             return;
@@ -452,11 +452,11 @@ public class InventoryM extends javax.swing.JPanel {
 
     private void UpdatedQtyKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_UpdatedQtyKeyReleased
         // TODO add your handling code here:
-        float qty, price;
+        double qty, price;
         double tot;
         try {
-            qty = Float.parseFloat(this.UpdatedQty.getText());
-            price = Float.parseFloat(this.AvgPrch.getText());
+            qty = Double.parseDouble(this.UpdatedQty.getText());
+            price = Double.parseDouble(this.AvgPrch.getText());
         } catch (NumberFormatException e) {
             e.printStackTrace();
             return;
@@ -483,8 +483,8 @@ public class InventoryM extends javax.swing.JPanel {
             String Qty = this.UpdatedQty.getText();
             String current = this.CurrentQty.getText();
 
-            float qtyNow = Float.parseFloat(Qty);
-            float qtyOld = Float.parseFloat(current);
+            double qtyNow = Double.parseDouble(Qty);
+            double qtyOld = Double.parseDouble(current);
 
             if (Qty.isBlank() && (current.equals(Qty))) {
                 JOptionPane.showMessageDialog(null, "Please enter the updated quantity", "Enter New Quantity", JOptionPane.ERROR_MESSAGE);
@@ -543,8 +543,8 @@ public class InventoryM extends javax.swing.JPanel {
             String Qty = this.UpdatedQty.getText();
             String current = this.CurrentQty.getText();
 
-            float qtyNow = Float.parseFloat(Qty);
-            float qtyOld = Float.parseFloat(current);
+            double qtyNow = Double.parseDouble(Qty);
+            double qtyOld = Double.parseDouble(current);
 
             if (Qty.isBlank() && (current.equals(Qty))) {
                 JOptionPane.showMessageDialog(null, "Please enter the updated quantity", "Enter New Quantity", JOptionPane.ERROR_MESSAGE);
