+++ src/PurchaseRegistration.java
@@ -179,7 +179,7 @@ public class PurchaseRegistration extends javax.swing.JFrame {
             ResultSet rs = select.executeQuery();
             if (rs.next()) {
                 int total = rs.getInt("Total");
-                float vat = rs.getFloat("Vat");
+                double vat = rs.getDouble("Vat");
 
                 DecimalFormat dfTotal = new DecimalFormat("#.##");
                 this.TotalAmount.setText(dfTotal.format(total));
@@ -976,15 +976,15 @@ public class PurchaseRegistration extends javax.swing.JFrame {
                         Bal.setString(1, BankPaid);
 
                         ResultSet rs = Bal.executeQuery();
-                        float IN, OUT, bal = 0;
+                        double IN, OUT, bal = 0;
                         if (rs.next()) {
-                            IN = rs.getFloat("BIN");
-                            OUT = rs.getFloat("BOUT");
+                            IN = rs.getDouble("BIN");
+                            OUT = rs.getDouble("BOUT");
                             bal = IN - OUT;
                         }
 
                         insert = con.prepareStatement("update bank set Balance=? where TxnId=?");
-                        insert.setFloat(1, bal);
+                        insert.setDouble(1, bal);
                         insert.setString(2, PurchaseID);
 
                         insert.executeUpdate();
@@ -1098,10 +1098,10 @@ public class PurchaseRegistration extends javax.swing.JFrame {
                     stock.executeUpdate();
 
                     CurrentStock currentStock = new CurrentStock();
-                    float current = currentStock.getCurrentStock(Codes)[0];
+                    double current = currentStock.getCurrentStock(Codes)[0];
 
                     Current = con.prepareStatement("update stock set CurrentStock=? where ItemCode=?");
-                    Current.setFloat(1, current);
+                    Current.setDouble(1, current);
                     Current.setString(2, Codes);
 
                     Current.executeUpdate();
@@ -1114,7 +1114,7 @@ public class PurchaseRegistration extends javax.swing.JFrame {
 
                     Current.executeUpdate();
 
-                    float amount = currentStock.getCurrentStock(Codes)[1];
+                    double amount = currentStock.getCurrentStock(Codes)[1];
 
                     Current = con.prepareStatement("update stock set  StockAmount=(CurrentStock*PurchasePrice) where ItemCode=?");
                     Current.setString(1, Codes);
@@ -1122,7 +1122,7 @@ public class PurchaseRegistration extends javax.swing.JFrame {
                     Current.executeUpdate();
 
                     Current = con.prepareStatement("update item set CurrentStock=? where ItemCode=?");
-                    Current.setFloat(1, current);
+                    Current.setDouble(1, current);
                     Current.setString(2, Codes);
 
                     Current.executeUpdate();
@@ -1260,13 +1260,13 @@ public class PurchaseRegistration extends javax.swing.JFrame {
         // TODO add your handling code here:
 
         int Qty = Integer.parseInt(this.SalesQty.getText());
-        float UnitPrice = Float.parseFloat(PurchaseRegistration.UnitPrice.getText());
-        float total;
+        double UnitPrice = Double.parseDouble(PurchaseRegistration.UnitPrice.getText());
+        double total;
         total = Qty * UnitPrice;
 
         String type = "B-18%";
         if (Tax.getSelectedItem().equals(type)) {
-            float vat = (float) (total * 0.18);
+            double vat = (double) (total * 0.18);
             DecimalFormat df = new DecimalFormat("#.##");
             VAT_Item.setText(df.format(vat));
         }
@@ -1279,13 +1279,13 @@ public class PurchaseRegistration extends javax.swing.JFrame {
         // TODO add your handling code here:
 
         int Qty = Integer.parseInt(this.SalesQty.getText());
-        float UnitPrice = Float.parseFloat(PurchaseRegistration.UnitPrice.getText());
-        float total;
+        double UnitPrice = Double.parseDouble(PurchaseRegistration.UnitPrice.getText());
+        double total;
         total = Qty * UnitPrice;
 
         String type = "B-18%";
         if (Tax.getSelectedItem().equals(type)) {
-            float vat = (float) (total * 0.18);
+            double vat = (double) (total * 0.18);
             DecimalFormat df = new DecimalFormat("#.##");
             VAT_Item.setText(df.format(vat));
         }
@@ -1352,10 +1352,10 @@ public class PurchaseRegistration extends javax.swing.JFrame {
                     insert.executeUpdate();
 
                     CurrentStock currentStock = new CurrentStock();
-                    float current = currentStock.getCurrentStock(id)[0];
+                    double current = currentStock.getCurrentStock(id)[0];
 
                     Current = con.prepareStatement("update stock set CurrentStock=? where ItemCode=?");
-                    Current.setFloat(1, current);
+                    Current.setDouble(1, current);
                     Current.setString(2, id);
 
                     Current.executeUpdate();
@@ -1368,7 +1368,7 @@ public class PurchaseRegistration extends javax.swing.JFrame {
 
                     Current.executeUpdate();
 
-                    float amount = currentStock.getCurrentStock(id)[1];
+                    double amount = currentStock.getCurrentStock(id)[1];
 
                     Current = con.prepareStatement("update stock set  StockAmount=(CurrentStock*PurchasePrice) where ItemCode=?");
                     Current.setString(1, id);
@@ -1376,7 +1376,7 @@ public class PurchaseRegistration extends javax.swing.JFrame {
                     Current.executeUpdate();
 
                     Current = con.prepareStatement("update item set CurrentStock=? where ItemCode=?");
-                    Current.setFloat(1, current);
+                    Current.setDouble(1, current);
                     Current.setString(2, id);
 
                     Current.executeUpdate();
@@ -1439,10 +1439,10 @@ public class PurchaseRegistration extends javax.swing.JFrame {
                         insert.executeUpdate();
 
                         CurrentStock currentStock = new CurrentStock();
-                        float current = currentStock.getCurrentStock(id)[0];
+                        double current = currentStock.getCurrentStock(id)[0];
 
                         Current = con.prepareStatement("update stock set CurrentStock=? where ItemCode=?");
-                        Current.setFloat(1, current);
+                        Current.setDouble(1, current);
                         Current.setString(2, id);
 
                         Current.executeUpdate();
@@ -1455,7 +1455,7 @@ public class PurchaseRegistration extends javax.swing.JFrame {
 
                         Current.executeUpdate();
 
-                        float amount = currentStock.getCurrentStock(id)[1];
+                        double amount = currentStock.getCurrentStock(id)[1];
 
                         Current = con.prepareStatement("update stock set  StockAmount=(CurrentStock*PurchasePrice) where ItemCode=?");
                         Current.setString(1, id);
@@ -1463,7 +1463,7 @@ public class PurchaseRegistration extends javax.swing.JFrame {
                         Current.executeUpdate();
 
                         Current = con.prepareStatement("update item set CurrentStock=? where ItemCode=?");
-                        Current.setFloat(1, current);
+                        Current.setDouble(1, current);
                         Current.setString(2, id);
 
                         Current.executeUpdate();
@@ -1512,13 +1512,13 @@ public class PurchaseRegistration extends javax.swing.JFrame {
         // TODO add your handling code here:
 
         int Qty = Integer.parseInt(this.SalesQty.getText());
-        float UnitPrice = Float.parseFloat(PurchaseRegistration.UnitPrice.getText());
-        float total;
+        double UnitPrice = Double.parseDouble(PurchaseRegistration.UnitPrice.getText());
+        double total;
         total = Qty * UnitPrice;
 
         String type = "B-18%";
         if (Tax.getSelectedItem().equals(type)) {
-            float vat = (float) (total * 0.18);
+            double vat = (double) (total * 0.18);
             DecimalFormat df = new DecimalFormat("#.##");
             VAT_Item.setText(df.format(vat));
         }
